requirements.txt
requests
pandas
python-dateutil
python-dotenv